pyproject.toml
[tool.poetry]
name = "prueba_git"
version = "0.1.0"
description = ""
authors = ["Monica Nataly <nata.moonk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.20"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
